Recommit "[libc++] NFC: Simplify substitutions by using lit recursive substitutions"
This re-commits e53cb56aaba0, which was reverted in 670a2856f08a because
it broke the LLVM `check-all` target. This commit addresses the underlying
issue by not setting the lit_config.recursiveExpansionLimit parameter of
the libc++ test suite, which is otherwise picked up by other test suites
in LLVM.

Once we've settled on a fix for the underlying issue with
lit_config.recursiveExpansionLimit, we can start using it
again in libc++, but for now we can just work around it.