[libc++] Port the time functions to z/OS
This patch adds a shim for missing time functions on z/OS, and adds a
layer of indirection to account for differences in the timespec struct
on different systems.

This was originally committed as af8e244a2612 and reverted in 39acd7f0dcce
because the original commit also checked-in unrelated changes.

Differential Revision: https://reviews.llvm.org/D87940