[libcxx] [test] Adjust separator form in fs.op.absolute for libc++ on windows
This test was previously tweaked in
82174e21e70519cf9b5a3b1878a5f04f5b768fec to match the output of
of MS STL (except that the MS STL fails on the testcase with an
empty path).

libc++ doesn't produce paths with all normalized separators (and the
spec doesn't mandate it to either).

Tweak the test reference to match exactly what libc++ produces. If
testing with a non-libc++ library, do a relaxed comparison that allows
the separators to differ.

Differential Revision: https://reviews.llvm.org/D98215