Revert "[libcxx][pretty printers] Import gdb module in gdb feature check"
This reverts commit e1e9c9f5ac4d6520ec2c2f44f3db6da3ab1b31d7.

This was my mistake. When gdb can find its data directory it'll
import it automatically. If it can't (like when you're using a
version from a build folder) you need to give it the data
directory path.

We're safe to assume gdb is installed for testing purposes
so it'll import it for us.