libcxx: Fix exception_prt abort when disable CONFIG_CXX_EXCEPTION.
The std::promise class contains a member variable of type exception_ptr (which is an exception type). If CONFIG_CXX_EXCEPTION is not enabled, using exception for checking will call abort in the exception_ptr class. Therefore, the implementation of the exception_ptr class needs to include a check to determine whether it is non-null before checking for an exception.

Signed-off-by: cuiziwei <cuiziwei@xiaomi.com>